added basic stats page
extracted shared function
