+++ public/index.html
@@ -14,7 +14,7 @@
     <script defer
             src="https://maps.googleapis.com/maps/api/js?key=AIzaSyDSz0lnzPJIFeWM7SpSARHmV-snwrAXd2s&libraries=geometry">
     </script>
-    <script src="locations-list-map.js"></script>
+    <script type="module" src="locations-list-map.js"></script>
     <title>#Findthemasks</title>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
