+++ public/locations-list-map.js
@@ -1,39 +1,4 @@
-function toDataByLocation(data) {
-  const headers = data.values[1];
-  const approvedIndex = headers.findIndex( e => e === 'approved' );
-  const stateIndex = headers.findIndex( e => e === 'state' );
-  const cityIndex = headers.findIndex( e => e === 'city' );
-  const data_by_location = {};
-
-  const published_entries = data.values.slice(1).filter((entry) => entry[approvedIndex] === "x");
-
-  published_entries.forEach( entry => {
-    const state = entry[stateIndex].trim();
-    const city = entry[cityIndex].trim();
-    let entry_array;
-    if (!(state in data_by_location) || !(city in data_by_location[state])) {
-      entry_array = [];
-      if (state in data_by_location) {
-        data_by_location[state][city] = entry_array;
-      } else {
-        data_by_location[state] = { [city]: entry_array };
-      }
-    } else {
-      entry_array = data_by_location[state][city];
-    }
-    const entry_obj = {};
-    headers.forEach( (value, index) => {
-      if (entry[index] !== undefined) {
-        entry_obj[value] = entry[index];
-      } else {
-        entry_obj[value] = "";
-      }
-    });
-    entry_array.push(entry_obj);
-  });
-
-  return data_by_location;
-}
+import toDataByLocation from './toDataByLocation.js';
 
 function createFiltersListHTML() {
   $('.locations-container').show();
