+++ public/stats.html
@@ -1,0 +1,54 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <!-- Global site tag (gtag.js) - Google Analytics -->
+  <script async src="https://www.googletagmanager.com/gtag/js?id=UA-161442455-1"></script>
+  <script>
+    window.dataLayer = window.dataLayer || [];
+    function gtag(){dataLayer.push(arguments);}
+    gtag('js', new Date());
+    gtag('config', 'UA-161442455-1');
+  </script>
+  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery.i18n/1.0.7/jquery.i18n.fallbacks.min.js"></script>
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery.i18n/1.0.7/jquery.i18n.min.js"></script>
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery.i18n/1.0.7/jquery.i18n.language.min.js"></script>
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery.i18n/1.0.7/jquery.i18n.messagestore.min.js"></script>
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery.i18n/1.0.7/jquery.i18n.parser.min.js"></script>
+  <script type="module" src="stats.js"></script>
+  <title>#Findthemasks</title>
+  <meta charset="UTF-8"/>
+  <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
+  <link href="https://fonts.googleapis.com/css?family=Roboto:400,700&display=swap" rel="stylesheet">
+  <meta name="description" content="Help #findthemasks for covid-19" />
+  <meta name="keywords" content="#findthemasks, masks, n95, coronavirus, covid-19" />
+  <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.4.1/css/bootstrap.min.css" integrity="sha384-Vkoo8x4CGsO3+Hhxv8T/Q5PaXtkKtu6ug5TOeNV6gBiFeWPGFN9MuhOf23Q9Ifjh" crossorigin="anonymous">
+  <style>
+
+  </style>
+
+  <!-- Make sure to translate these meta tags when creating a new language version -->
+
+  <!-- English -->
+  <meta lang="en" property="og:title" content="#findthemasks"/>
+  <meta lang="en" property="og:url" content="https://findthemasks.com/"/>
+  <meta lang="en" property="og:description" content="America’s frontline healthcare workers are treating COVID-19 patients without adequate protective gear, risking their lives! We need to find the masks. All of these masks can save lives now if you get them into the hands of healthcare workers."/>
+
+  <!-- French -->
+  <meta lang="fr" property="og:title" content="#findthemasks"/>
+  <meta lang="fr" property="og:url" content="https://findthemasks.com/"/>
+  <meta lang="fr" property="og:description" content="Les travailleurs de la santé de première ligne du monde entier traitent les patients atteints de COVID-19 sans équipement de protection adéquat, risquant leur vie! Nous devons trouver les masques. Tous ces masques peuvent maintenant sauver des vies si vous les remettez entre les mains de professionnels de la santé."/>
+
+  <meta property="og:image" content="https://findthemasks.com/images/n95.jpg"/>
+</head>
+
+<body>
+<main class="content">
+  <div class="container" style="padding-bottom: 40px;">
+    <h2 style="padding: 30px 0px 20px 0px;" class='i18n' data-i18n="ftm-stats">Stats</h2>
+
+    <div class="stats-container"></div>
+  </div>
+</main>
+</body>
+</html>
